backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv